Move StringBuilderProcessOutputLineListener out of internal so Eclipse can use it
